Add 1-for-1 signal substitution optimization
This optimization looks for constraints of the form:
   0 * 0 = ax + by + c
or
   0 * 0 = ax + c
and uses them to eliminate x.

I've implemented it *very* inefficiently, so it only works on small
circuits right now.

On poseidon, it reduces the constraint count from 1981 to 705 (2.5x
improvement!). I suspect it will do even better on larger (specifically,
more hierarchical) circuits.