Use a fresh variable for each guard condition.
That is, when you use Circify.pushGuard term, a new variable is
introduced and assigned to term. This seems to prevent term size blowup
for reasons I don't fully understand. There must be some kind of
term-guard cyclic inclusion which can lead to unbounded growth. At any
rate, the fix is simple and obviously sound.